% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estM.R
\name{estM}
\alias{estM}
\title{Estimate the resample size M from N sample}
\usage{
estM(N, p, lambda = 0.025)
}
\arguments{
\item{N}{A number of total sample size}

\item{p}{A number of degree of nonregularity}

\item{lambda}{A number between 0 and 1 that governs the smallest acceptable resample size. Typical choice are between [0.025, 0.1], default value is 0.025}
}
\value{
\item{M:}{ An estimated resample size}
}
\description{
A function that estimate the resample size M from N sample with given degree of nonregularity in the stage and a prespecified parameter \code{lambda}
}
\examples{
M_hat <- estM(N = 100, p = 0.45, lambda = 0.02) 
M_hat

}
