% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interact.terms.R
\name{interact.terms}
\alias{interact.terms}
\title{Identify interaction terms}
\usage{
interact.terms(formula, target_covariate)
}
\arguments{
\item{formula}{A formula that specifying the regression model}

\item{target_covariate}{A string that indicating the name of the target covariate}
}
\value{
A vector of covariate names that interact with the target covariate
}
\description{
A function that identifying all interaction covariates with a target covariate in a model formula
}
\examples{
model_formula <- formula(Y ~ A1 * X2 + A1:X3 + X4)
target_covariate <- "A1"
interacting_variables <- interact.terms(model_formula, target_covariate)

}
