% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoN.R
\name{MoN}
\alias{MoN}
\title{Inference under nonregrularity with M-out-of-N bootstrap}
\usage{
MoN(
  completeData,
  s1Formula,
  s2Formula,
  s2Treat,
  cluster,
  bootNum = 1000,
  alpha = 0.05,
  nu = 0.05,
  lambda = 0.025,
  ...
)
}
\arguments{
\item{completeData}{A matrix contains all data including covariates, outcome and treatment}

\item{s1Formula}{A formula that specifying the regression model for stage 1 Q-function}

\item{s2Formula}{A formula that specifying the regression model for stage 2 Q-function}

\item{s2Treat}{A string that records variable name of stage 2 treatment in s2_data}

\item{cluster}{A string that records variable name of the cluster identifier}

\item{bootNum}{A number indicates number of bootstrap}

\item{alpha}{A number indicates the type I error of inference, the default value is 0.05}

\item{nu}{A number indicates the threshold for T-statistics, the default value is 0.001}

\item{lambda}{A number between 0 and 1 that governs the smallest acceptable resample size. Typical choice are between [0.025, 0.1], default value is 0.025}

\item{...}{Additional parameters in lm or lme}
}
\value{
A list containing 
\itemize{
  \item{s1Coefficients}{Stage 1 regression coefficients}
  \item{s2Coefficients}{Stage 2 regression coefficients}
  \item{s1Inference}{Stage 1 coefficients and confidence interval}
  \item{s2Inference}{Stage 2 coefficients and confidence interval}
  \item{estM}{Estimated stage 1 bootstrap resample size}
}
}
\description{
A function that using M resample size out of N samples to conduct bootstrap and make inference on regression parameters to adjust for nonregrularity
}
